\name{frab-package}
\alias{frab-package}
\docType{package}
\title{\packageTitle{frab}}
\description{\packageDescription{frab}}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{frab}
\packageIndices{frab}
}
\author{
\packageAuthor{frab}

Maintainer: \packageMaintainer{frab}
}
\keyword{ package }
\examples{
x <- frab(c(a=1, b=2, c=5))
y <- frab(c(b=-2, c=1, d=8))

x+y

}
