\name{Compare-methods}
\docType{methods}
\alias{Compare-methods}
\alias{Compare,frab,frab-method}
\alias{Compare,frab,numeric-method}
\alias{Compare,numeric,frab-method}
\alias{frab_eq}
\alias{frab_compare_frab}
\alias{frab_eq_num}
\alias{frab_gt_num}
\alias{frab_ge_num}
\alias{frab_lt_num}
\alias{frab_le_num}
\alias{frab_compare_numeric}
\alias{num_eq_frab}
\alias{num_gt_frab}
\alias{num_ge_frab}
\alias{num_lt_frab}
\alias{num_le_frab}
\alias{numeric_compare_frab}
\alias{frab_compare_error}
\title{Comparision methods}
\description{
  Methods for comparision (greater than, etc) in the frab package
}
\note{

note here  
}
\examples{
rfrab() < 5
a <- rfrab(26,sym=letters)
a[a<4] <- 100
  }
\keyword{methods}
\keyword{math}
